pyproject.toml
[tool.poetry]
name = "fermigier-com"
version = "0.1.0"
description = ""
authors = ["Stefane Fermigier <sf@abilian.com>"]
readme = "README.md"
packages = [{include = "fermigier_com", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
Jinja2 = "^3.1.2"
Markdown = "^3.4.1"
typer = "^0.7.0"
attrs = "^22.1.0"
pyyaml = "^6.0"
argh = "^0.26.2"
watchdog = {extras = ["watchmedo"], version = "^2.2.0"}
honcho = "^1.1.0"
watchfiles = "^0.18.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
devtools = "^0.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
alamano = "fermigier_com.cli:app"
